+++ p2panda-rs/src/test_utils/mod.rs
@@ -2,20 +2,16 @@
 
 //! # p2panda test-utils
 //!
-//! This module provides tools used for testing and generating test data for `p2panda-rs` and `p2panda-js`. 
+//! This module provides tools used which can be used for testing and generating test data for `p2panda-rs` and `p2panda-js`. 
 //!
 //! It includes:
-//! - a mock node
-//! - a mock client
+//! - fixtures and templates which can be injected into tests
+//! - mock node and client with experimental materialisation logic implemented
 //! - methods for generating test data (used in `p2panda-js` tests)
 
-#[cfg(test)]
 pub mod fixtures;
-pub mod logs;
-pub mod node;
-pub mod materialiser;
-pub mod data;
-pub mod client;
+pub mod test_data;
+pub mod mocks;
 mod utils;
 
 pub use utils::*;
