+++ p2panda-rs/src/test_utils/test_data/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+//! Structs and methods for generating test data form p2panda.
+
+pub mod json_data;
